Rename and consolidate per pr.
