--- src/app/azureResources/components/azureResourcesContainer.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-import { RouteComponentProps } from 'react-router-dom';
-import { ConnectionStringsViewContainer } from '../../connectionStrings/components/connectionStringsViewContainer';
-
-export const AzureResourcesContainer: React.FC<RouteComponentProps> = props => {
-       return <ConnectionStringsViewContainer {...props} />;
-};
