+++ src/app/connectionStrings/components/connectionStringsView.tsx
@@ -3,12 +3,19 @@
  * Licensed under the MIT License
  **********************************************************/
 import * as React from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { RouteComponentProps } from 'react-router-dom';
 import { CommandBar } from 'office-ui-fabric-react/lib/CommandBar';
 import { ConnectionString  } from './connectionString';
 import { ConnectionStringEditView } from './connectionStringEditView';
 import { useLocalizationContext } from '../../shared/contexts/localizationContext';
 import { ResourceKeys } from '../../../localization/resourceKeys';
 import { CONNECTION_STRING_LIST_MAX_LENGTH } from '../../constants/browserStorage';
+import { StateInterface } from '../../shared/redux/state';
+import { upsertConnectionStringAction, deleteConnectionStringAction, setConnectionStringsAction } from '../actions';
+import { setActiveAzureResourceByConnectionStringAction } from '../../azureResource/actions';
+import { ROUTE_PARTS } from '../../constants/routes';
+import { formatConnectionStrings } from '../../shared/utils/hubConnectionStringHelper';
 import '../../css/_layouts.scss';
 import './connectionStringsView.scss';
 
@@ -79,5 +86,39 @@ export const ConnectionStringsView: React.FC<ConnectionStringsViewProps> = props
                 />
             }
         </div>
+    );
+};
+
+export const ConnectionStringsViewContainer: React.FC<RouteComponentProps> = props => {
+    const connectionStrings = useSelector((state: StateInterface) => state.connectionStringsState.connectionStrings);
+    const dispatch = useDispatch();
+
+    const onUpsertConnectionString = (newConnectionString: string, connectionString?: string) => {
+        dispatch(upsertConnectionStringAction({newConnectionString, connectionString}));
+    };
+
+    const onDeleteConnectionString = (connectionString: string) => {
+        dispatch(deleteConnectionStringAction(connectionString));
+    };
+
+    const onSelectConnectionString = (connectionString: string, hostName: string) => {
+        const updatedConnectionStrings = formatConnectionStrings(connectionStrings, connectionString);
+
+        dispatch(setConnectionStringsAction(updatedConnectionStrings));
+        dispatch(setActiveAzureResourceByConnectionStringAction({
+            connectionString,
+            hostName
+        }));
+
+        props.history.push(`/${ROUTE_PARTS.RESOURCE}/${hostName}/${ROUTE_PARTS.DEVICES}`);
+    };
+
+    return (
+        <ConnectionStringsView
+            onUpsertConnectionString={onUpsertConnectionString}
+            onDeleteConnectionString={onDeleteConnectionString}
+            onSelectConnectionString={onSelectConnectionString}
+            connectionStrings={connectionStrings}
+        />
     );
 };
